- refactoring - add test for voting

+++ src/main/java/ru/ilin/restvote/service/VotingService.java
@@ -5,7 +5,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.ilin.restvote.model.Vote;
 import ru.ilin.restvote.repository.VoteRepository;
+import ru.ilin.restvote.utils.exception.IllegalRequestDataException;
 
+import java.time.LocalDateTime;
+
+import static ru.ilin.restvote.utils.DateTimeUtil.TIME_FORMATTER;
 import static ru.ilin.restvote.utils.validation.ValidationUtil.checkNotFoundWithId;
 
 @Service
@@ -30,7 +34,10 @@ public class VotingService {
     }
 
     @Transactional
-    public Vote createVote(int restaurantId, int userId) {
+    public Vote createVote(int restaurantId, int userId, LocalDateTime dateTineVoting) {
+        if (dateTineVoting.toLocalTime().isAfter(Vote.VOTE_LAST_TIME_FOR_VOTING)) {
+            throw new IllegalRequestDataException("Vote can only be changed before " + Vote.VOTE_LAST_TIME_FOR_VOTING.format(TIME_FORMATTER));
+        }
         repository.deleteByUserId(userId);
         return repository.save(new Vote(restaurantService.get(restaurantId), userService.get(userId)));
     }
